services/authService/requirements.txt
alembic==1.16.5
asyncpg==0.30.0
bcrypt==4.2.0
email-validator==2.2.0
fastapi==0.116.1
httpx==0.28.1
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
redis==5.2.1
SQLAlchemy==2.0.43
uvicorn[standard]==0.35.0
